+++ tests/unit/test_utilities.py
@@ -13,7 +13,9 @@ from harmony_service_lib.util import config
 from hoss.exceptions import (
     CustomError,
     CustomNoRetryError,
+    StagingFailed,
     UrlAccessFailed,
+    UrlAccessFailedWithNoRetries,
     UrlAccessForbidden,
 )
 from hoss.harmony_log_context import set_logger
@@ -37,6 +39,12 @@ class TestUtilities(TestCase):
     @classmethod
     def setUpClass(cls):
         cls.harmony_500_error = ServerException('I can\'t do that')
+        cls.harmony_500_status_code_error = ServerException(
+            'Unable to download due to status code: 500 and content'
+        )
+        cls.harmony_400_status_code_error = ServerException(
+            'Unable to download due to status code: 400 and content'
+        )
         cls.harmony_auth_error = ForbiddenException('You can\'t do that.')
         cls.config = config(validate=False)
         cls.logger = getLogger('test')
@@ -129,6 +137,28 @@ class TestUtilities(TestCase):
             )
             mock_util_download.reset_mock()
 
+        with self.subTest('Harmony server exception 500 error should be retried.'):
+            mock_util_download.side_effect = [
+                self.harmony_500_status_code_error,
+                http_response,
+            ]
+
+            with self.assertRaises(UrlAccessFailed) as context:
+                download_url(test_url, output_directory, access_token, self.config)
+
+            self.assertIsInstance(context.exception, CustomError)
+            self.assertNotIsInstance(context.exception, CustomNoRetryError)
+
+            mock_util_download.assert_called_once_with(
+                test_url,
+                output_directory,
+                self.logger,
+                access_token=access_token,
+                data=None,
+                cfg=self.config,
+            )
+            mock_util_download.reset_mock()
+
         with self.subTest('A 403 error (forbidden) is not retried.'):
             mock_util_download.side_effect = [self.harmony_auth_error, http_response]
 
@@ -136,6 +166,28 @@ class TestUtilities(TestCase):
                 download_url(test_url, output_directory, access_token, self.config)
 
             self.assertIsInstance(context.exception, CustomNoRetryError)
+
+            mock_util_download.assert_called_once_with(
+                test_url,
+                output_directory,
+                self.logger,
+                access_token=access_token,
+                data=None,
+                cfg=self.config,
+            )
+            mock_util_download.reset_mock()
+
+        with self.subTest('A 400 error (Bad Request) is not retried.'):
+            mock_util_download.side_effect = [
+                self.harmony_400_status_code_error,
+                http_response,
+            ]
+
+            with self.assertRaises(UrlAccessFailedWithNoRetries) as context:
+                download_url(test_url, output_directory, access_token, self.config)
+
+            self.assertIsInstance(context.exception, CustomNoRetryError)
+            self.assertNotIsInstance(context.exception, CustomError)
 
             mock_util_download.assert_called_once_with(
                 test_url,
@@ -390,6 +442,19 @@ class TestUtilities(TestCase):
 
         with self.subTest('UrlAccessFailed (retryable) raises HarmonyException.'):
             failed_exception = UrlAccessFailed(test_url, 500)
+            with self.assertRaises(HarmonyException) as context:
+                raise_from_hoss_exception(failed_exception)
+            self.assertNotIsInstance(context.exception, NoRetryException)
+
+        with self.subTest(
+            'UrlAccessFailedWithNoRetries (no-retry) raises NoRetryException.'
+        ):
+            failed_exception = UrlAccessFailedWithNoRetries(test_url, 400)
+            with self.assertRaises(NoRetryException):
+                raise_from_hoss_exception(failed_exception)
+
+        with self.subTest('StagingFailed (retryable) raises HarmonyException.'):
+            failed_exception = StagingFailed('Staging failed, Connection timeout')
             with self.assertRaises(HarmonyException) as context:
                 raise_from_hoss_exception(failed_exception)
             self.assertNotIsInstance(context.exception, NoRetryException)